Updated gradle, scripts, moved deps to catalog
* Updated gradle wrapper version to 8.7
  * Fixed build scripts
* Updated variuous dependencies
  * Updated lavaplayer to fix version
  * Updated groovy, jackson, swagger, etc. to latest patch version
* Moved all dependencies to catalog at gradle/libs.versions.toml
* Updated dependencies for all modules
