+++ .gitignore
@@ -23,6 +23,11 @@ hs_err_pid*
 
 # Ignore Gradle project-specific cache directory
 .gradle
+gradle.properties
 
 # Ignore Gradle build output directory
 build
+
+# Custom
+.idea
+
