+++ .idea/jarRepositories.xml
@@ -36,5 +36,15 @@
       <option name="name" value="MavenLocal" />
       <option name="url" value="file:/$MAVEN_REPOSITORY$/" />
     </remote-repository>
+    <remote-repository>
+      <option name="id" value="arbjergDevSnapshots" />
+      <option name="name" value="arbjergDevSnapshots" />
+      <option name="url" value="https://maven.lavalink.dev/snapshots" />
+    </remote-repository>
+    <remote-repository>
+      <option name="id" value="Gradle Central Plugin Repository" />
+      <option name="name" value="Gradle Central Plugin Repository" />
+      <option name="url" value="https://plugins.gradle.org/m2" />
+    </remote-repository>
   </component>
 </project>
