+++ audio/build.gradle
@@ -1,12 +1,13 @@
 plugins {
-    id 'com.asm.tavern.discord.groovy-library-conventions'
+    id 'java-library'
+    id 'com.tavern.groovy-common-conventions'
 }
 
 dependencies {
     implementation project(':domain')
     implementation project(':utilities')
-    // Repo abandoned: implementation 'com.sedmelluq:lavaplayer:1.3.61'
-    // Using for (walkyst) instead
-    implementation "dev.arbjerg:lavaplayer:6d14c5b2d96fd953c671cfae8f9b9ea11aa20d5a-SNAPSHOT"
-}
 
+    implementation libs.bundles.core
+    implementation libs.lavaplayer
+    implementation libs.discordJda
+}
