--- buildSrc/src/main/groovy/com.asm.tavern.discord.groovy-application-conventions.gradle
@@ -1,19 +1,0 @@
-/*
- * This file was generated by the Gradle 'init' task.
- */
-
-plugins {
-    // Apply the common convention plugin for shared build configuration between library and application projects.
-    id 'com.asm.tavern.discord.groovy-common-conventions'
-
-    // Apply the application plugin to add support for building a CLI application in Java.
-    id 'application'
-}
-
-dependencies {
-	// Logger implementation
-    runtime group: 'org.apache.logging.log4j', name: 'log4j-slf4j-impl', version: '2.13.3'
-    runtime group: 'org.apache.logging.log4j', name: 'log4j-1.2-api', version: '2.14.0'
-    runtime group: 'org.apache.logging.log4j', name: 'log4j-api', version: '2.13.3'
-    runtime group: 'org.apache.logging.log4j', name: 'log4j-core', version: '2.13.3'
-}
