--- buildSrc/src/main/groovy/com.asm.tavern.discord.groovy-common-conventions.gradle
@@ -1,66 +1,0 @@
-/*
- * This file was generated by the Gradle 'init' task.
- */
-
-plugins {
-    // Apply the groovy Plugin to add support for Groovy.
-    id 'groovy'
-}
-
-repositories {
-    // Use JCenter for resolving dependencies.
-    jcenter()
-    mavenCentral()
-    // Has to add the below for lavaplayer fork
-    mavenLocal()
-    maven {
-        url 'https://jitpack.io'
-    }
-    maven {
-        url = 'https://m2.dv8tion.net/releases'
-    }
-    maven {
-        name "arbjergDevSnapshots"
-        url "https://maven.lavalink.dev/snapshots"
-    }
-}
-
-dependencies {
-    // Use the latest Groovy version for building this library
-    // implementation 'org.codehaus.groovy:groovy-all'
-    compile "org.codehaus.groovy:groovy:3.0.7"
-
-    compile group: 'org.springframework', name: 'spring-context', version: '5.3.3'
-    compile group: 'org.springframework', name: 'spring-core', version: '5.3.3'
-    compile group: 'org.springframework', name: 'spring-beans', version: '5.3.3'
-
-    // Discord JDS API
-    compile 'net.dv8tion:JDA:5.0.0-beta.6'
-
-    // Logging
-    compile group: 'org.slf4j', name: 'slf4j-simple', version: '2.0.7'
-    compile group: 'org.slf4j', name: 'slf4j-api', version: '2.0.7'
-    compile group: 'org.slf4j', name: 'slf4j-ext', version: '2.0.7'
-
-    // Util
-    compile group: 'com.google.guava', name: 'guava', version: '30.1-jre'
-
-    // mandatory dependencies for using Spock
-    testCompile platform("org.spockframework:spock-bom:2.0-M4-groovy-3.0")
-    testCompile "org.spockframework:spock-core"
-    testCompile "org.spockframework:spock-junit4"  // you can remove this if your code does not rely on old JUnit 4 rules
-
-    // optional dependencies for using Spock
-    testCompile "org.hamcrest:hamcrest-core:2.2"   // only necessary if Hamcrest matchers are used
-    testRuntime "net.bytebuddy:byte-buddy:1.10.10" // allows mocking of classes (in addition to interfaces)
-    testRuntime "org.objenesis:objenesis:3.1"      // allows mocking of classes without default constructor (together with ByteBuddy or CGLIB)
-
-}
-
-// tasks.named('test') {
-test {
-    useJUnitPlatform()
-    testLogging {
-        events "passed", "skipped", "failed"
-    }
-}
