+++ buildSrc/src/main/groovy/com.tavern.groovy-common-conventions.gradle
@@ -1,0 +1,33 @@
+/*
+ * Common Plugin for all modules
+ */
+
+plugins {
+    // Apply the groovy Plugin to add support for Groovy.
+    id 'groovy'
+}
+
+repositories {
+    // Use JCenter for resolving dependencies.
+    jcenter()
+    mavenCentral()
+    // Has to add the below for lavaplayer fork
+    mavenLocal()
+    maven {
+        url 'https://jitpack.io'
+    }
+    maven {
+        url = 'https://m2.dv8tion.net/releases'
+    }
+    maven {
+        name "arbjergDevSnapshots"
+        url "https://maven.lavalink.dev/snapshots"
+    }
+}
+
+test {
+    useJUnitPlatform()
+    testLogging {
+        events "passed", "skipped", "failed"
+    }
+}
