+++ discord/src/main/groovy/com/asm/tavern/discord/discord/audio/SongsCommandHandler.groovy
@@ -1,6 +1,6 @@
 package com.asm.tavern.discord.discord.audio
 
-import com.asm.tavern.discord.utilities.DiscordUtils
+import com.asm.tavern.discord.discord.DiscordUtils
 import com.asm.tavern.domain.model.TavernCommands
 import com.asm.tavern.domain.model.audio.Song
 import com.asm.tavern.domain.model.audio.SongService
