+++ domain/build.gradle
@@ -1,9 +1,15 @@
 plugins {
-    id 'com.asm.tavern.discord.groovy-library-conventions'
+    id 'java-library'
+    id 'com.tavern.groovy-common-conventions'
 }
 
 dependencies {
     implementation project(':utilities')
     implementation 'org.json:json:20231013'
-    implementation "dev.arbjerg:lavaplayer:6d14c5b2d96fd953c671cfae8f9b9ea11aa20d5a-SNAPSHOT"
+
+    implementation libs.bundles.core
+    // TODO: EMM Remove this dependency from domain
+    implementation libs.lavaplayer
+    // TODO: EMM Remove this dependency from domain
+    implementation libs.discordJda
 }
