+++ gradle.properties
@@ -1,0 +1,2 @@
+org.gradle.java.home=C:/Program Files/Eclipse Adoptium/jdk-11.0.22.7-hotspot
+
