+++ repository/build.gradle
@@ -1,13 +1,12 @@
 plugins {
-    id 'com.asm.tavern.discord.groovy-library-conventions'
+    id 'java-library'
+    id 'com.tavern.groovy-common-conventions'
 }
 
 dependencies {
     implementation project(':utilities')
     implementation project(':domain')
 
-    compile group: 'com.fasterxml.jackson.core', name: 'jackson-core', version: '2.12.1'
-    compile group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.12.1'
-    compile group: 'com.fasterxml.jackson.core', name: 'jackson-annotations', version: '2.12.1'
-    compile group: 'com.fasterxml.jackson.datatype', name: 'jackson-datatype-jdk8', version: '2.12.1'
+    implementation libs.bundles.core
+    implementation libs.bundles.jackson
 }
