+++ utilities/build.gradle
@@ -1,3 +1,10 @@
 plugins {
-    id 'com.asm.tavern.discord.groovy-library-conventions'
+    id 'java-library'
+    id 'com.tavern.groovy-common-conventions'
+}
+
+dependencies {
+    implementation libs.bundles.core
+
+    testImplementation libs.bundles.spock
 }
